feat: 1st iteration progress

+++ src/wishes/wishes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Req } from '@nestjs/common';
+import { ForbiddenException, Injectable, Req } from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -36,28 +36,96 @@ export class WishesService {
     });
   }
 
-  update(id: number, updateWishDto: UpdateWishDto) {
-    this.wishRepository.update(id, updateWishDto);
-  }
-
-  findTopOne() {
-    return this.wishRepository.find({
-      take: 1,
-      order: { copied: 'DESC' },
+  async update(id: number, updateWishDto: UpdateWishDto, user: User) {
+    const wish = await this.wishRepository.findOne({
+      where: { id },
       relations: {
         owner: true,
         offers: true,
       },
     });
+
+    if (wish.owner.id !== user.id) {
+      throw new ForbiddenException('Нельзя изменять чужие желания');
+    }
+
+    if (wish.raised > 0) {
+      throw new ForbiddenException(
+        'Нельзя менять стоимость желания, на которое уже скинулись',
+      );
+    }
+  }
+
+  findTopOne() {
+    return this.wishRepository.find({
+      order: { copied: 'DESC' },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+      take: 1,
+    });
   }
 
   findLastOne() {
     return this.wishRepository.find({
-      take: 1,
       order: { createdAt: 'DESC' },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+      take: 1,
     });
   }
-  remove(id: number) {
+  findUserWishes(user: User) {
+    return this.wishRepository.findOne({
+      where: { owner: { id: user.id } },
+    });
+  }
+  async remove(id: number, user: User) {
+    const wish = await this.wishRepository.findOne({
+      where: { id },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+    });
+    if (wish.owner.id !== user.id) {
+      throw new ForbiddenException('Нельзя удалять чужие желания');
+    }
     return this.wishRepository.delete({ id });
+  }
+
+  async copy(wish: Wish, user: User) {
+    if (user.id === wish.owner.id) {
+      throw new ForbiddenException('Нельзя копировать свой подарок');
+    }
+
+    const existingCopy = await this.wishRepository.findOne({
+      where: { name: wish.name, owner: { id: user.id } },
+    });
+
+    if (existingCopy) {
+      throw new ForbiddenException('Копия уже существует');
+    }
+
+    const copiedWish = this.wishRepository.create({
+      name: wish.name,
+      image: wish.image,
+      link: wish.link,
+      price: wish.price,
+      description: wish.description,
+      owner: user,
+    });
+
+    await this.wishRepository.update(wish.id, { copied: wish.copied + 1 });
+
+    return {
+      name: copiedWish.name,
+      image: copiedWish.image,
+      link: copiedWish.link,
+      price: copiedWish.price,
+      description: copiedWish.description,
+    };
   }
 }
